refactor: lib for test solution
add lib for executing test
