+++ Cargo.toml
@@ -2,7 +2,7 @@
 members = [
     "problems",
     "rust/library",
-    "rust/solution"
+    "rust/test_executor"
 ]
 
 [package]
@@ -16,10 +16,11 @@ readme = "README.md"
 
 [[test]]
 name = "solution_test"
-path = "rust/solution/test/test.rs"
+path = "rust/test_executor/src/test.rs"
 
 [dependencies]
 serde_json = "1.0"
 rand = "0.8.4"
 regex = "1.10.5"
 problems = { path = "problems", features = ["solution"] }
+test_executor = { path = "rust/test_executor", features = ["timer", "test_case"] }
