+++ rust/library/src/lib/list_node.rs
@@ -15,21 +15,22 @@ impl ListNode {
     }
 }
 
-pub fn int_array_to_list_node(arr: Vec<i32>) -> Option<Box<ListNode>> {
+pub fn int_array_to_list_node(arr: &Vec<i32>) -> Option<Box<ListNode>> {
     let mut dummy = Some(Box::new(ListNode::new(0)));
     let mut p = dummy.as_mut().unwrap();
     for i in arr {
-        p.next = Some(Box::new(ListNode::new(i)));
+        p.next = Some(Box::new(ListNode::new(*i)));
         p = p.next.as_mut().unwrap();
     }
     dummy.unwrap().next
 }
 
-pub fn list_node_to_int_array(mut head: Option<Box<ListNode>>) -> Vec<i32> {
+pub fn list_node_to_int_array(head: &Option<Box<ListNode>>) -> Vec<i32> {
     let mut res = vec![];
-    while let Some(node) = head {
-        res.push(node.val);
-        head = node.next;
+    let mut node = head.clone();
+    while let Some(n) = node {
+        res.push(n.val);
+        node = n.next;
     }
     res
 }
