+++ rust/library/src/lib/tree_node.rs
@@ -21,7 +21,7 @@ impl TreeNode {
     }
 }
 
-pub fn array_to_tree(arr: Vec<Option<i32>>) -> Option<Rc<RefCell<TreeNode>>> {
+pub fn array_to_tree(arr: &Vec<Option<i32>>) -> Option<Rc<RefCell<TreeNode>>> {
     if arr.is_empty() {
         return None;
     }
