+++ rust/test_executor/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "solution"
+name = "test_executor"
 version = "0.1.0"
 edition = "2021"
 rust-version = "1.79.0"
@@ -7,10 +7,10 @@ authors = ["benhao"]
 description = "LeetCode Rust Test Main"
 readme = "../README.md"
 
+[features]
+timer = []
+test_case = []
+
 [dependencies]
 serde_json = "1.0"
 problems = { path = "../../problems", features = ["solution"] }
-
-[[test]]
-name = "solution_test"
-path = "test/test.rs"
