+++ rust/test_executor/src/lib.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "timer")]
+pub mod timer;
+
+#[cfg(feature = "test_case")]
+pub mod test_case;
